+++ crates/biome_js_analyze/tests/specs/style/useImportType/valid-unused-react-types.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+
+function Component() {
+    const onClick = (event: React.MouseEvent) => { };
+
+    return <div onClick={onClick}></div>;
+}